+++ example/android/app/src/main/kotlin/com/example/sticker_keyboard_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_social_keyboard_example
+package com.example.sticker_keyboard_example
 
 import io.flutter.embedding.android.FlutterActivity