+++ lib/sticker_keyboard.dart
@@ -1,0 +1,7 @@
+library sticker_keyboard;
+
+export 'package:sticker_keyboard/src/sticker_keyboard.dart';
+export 'package:sticker_keyboard/models/category_sticker.dart';
+export 'package:sticker_keyboard/models/keyboard_config.dart';
+export 'package:sticker_keyboard/models/sticker.dart';
+export 'package:sticker_keyboard/utils/sticker_picker_utils.dart';