+++ lib/widgets/search/sticker_search.dart
@@ -1,9 +1,10 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter_social_keyboard/models/keyboard_config.dart';
-import 'package:flutter_social_keyboard/models/sticker.dart';
-import 'package:flutter_social_keyboard/utils/sticker_picker_internal_utils.dart';
-import 'package:flutter_social_keyboard/utils/sticker_picker_utils.dart';
+import 'package:sticker_keyboard/models/keyboard_config.dart';
+import 'package:sticker_keyboard/models/sticker.dart';
+import 'package:sticker_keyboard/utils/sticker_picker_internal_utils.dart';
+import 'package:sticker_keyboard/utils/sticker_picker_utils.dart';
 
 class StickerSearch extends StatefulWidget {
   const StickerSearch({
@@ -133,7 +134,7 @@ class Calculates extends State<StickerSearch> {
               padding: widget.keyboardConfig.gridPadding,
               crossAxisCount: widget.keyboardConfig.stickerColumns,
               mainAxisSpacing: widget.keyboardConfig.stickerVerticalSpacing,
-              crossAxisSpacing: widget.keyboardConfig.emojiHorizontalSpacing,
+              crossAxisSpacing: widget.keyboardConfig.stickerHorizontalSpacing,
               children: [
                 for (int i = 0; i < stickers.length; i++)
                   GestureDetector(
@@ -146,12 +147,21 @@ class Calculates extends State<StickerSearch> {
                         );
                       }
                     },
-                    child: Image.asset(
-                      stickers[i].assetUrl,
-                      errorBuilder: ((context, error, stackTrace) =>
-                          const Icon(Icons.error)),
-                      fit: BoxFit.cover,
-                    ),
+                    child: stickers[i].assetUrl.startsWith('http')
+                        ? CachedNetworkImage(
+                            imageUrl: stickers[i].assetUrl,
+                            placeholder: (context, url) =>
+                                const CircularProgressIndicator.adaptive(),
+                            errorWidget: (context, url, error) =>
+                                const Icon(Icons.error),
+                            fit: BoxFit.cover,
+                          )
+                        : Image.asset(
+                            stickers[i].assetUrl,
+                            errorBuilder: ((context, error, stackTrace) =>
+                                const Icon(Icons.error)),
+                            fit: BoxFit.cover,
+                          ),
                   ),
               ],
             ),